requirements.txt
keyboard==0.13.5
mss==9.0.1
Pillow==10.4.0
requests==2.31.0
pytesseract==0.3.13
